First pass at debranding
This should update to OpenVox in help output and man pages. It also adds
a debranded README and gemspec.

If you'd like to help, feel free to check out this branch/PR and push
your own commits. Stay within scope though, we're focusing first on the
man pages and any user interaction. We'll get to docs later, that's
gonna be a beast.

If you'd like to help, then the easiest way to do so is to check out
this branch/PR and directly add your own commits.

Guidelines
- Do not change any command names.
- *Do not change anything functional!!*
- If any command output includes the name of the tool, change it as
  appropriate or add something like "provided by OpenVox"
    - if the name is extraneous, then just remove it. For example in the
      message "cannot find puppet server" I just removed the word
      "puppet".
- When there is a reference to 'Puppet' that means 'the Puppet language'
  then change it to 'the Puppet language' to be more precise. (OpenVox
  implements the Puppet language. There's no such thing as an OpenVox
  language.)
- Do minor content editing for readability as appropriate.
- We are currently focusing only on user experience and branding. It
  should be very clear to a user that they're running the OpenVox
  implementation of Puppet. It's not yet important that all the internal
  and developer documentation or comments reflect this.